Apply default button margin only in combination with auto width
Buttons apply a default right margin of 0.5rem. It's assumed this is intended to create spacing between a set of buttons (citation needed). Since a button occupies full width at smaller viewport sizes, the margin is unnecessary. In most cases, this is not a problem due to details around how specifications define that widths be defined. However, there are a few scenarios where it could be an issue:

1. When overriding the default width of the button
2. If applicable, using RTL direction instead of LTR

Regardless how common the above scenarios are, based on the expected usage of the margin, the expected usefulness would be limited to instances where a button is _not_ full width.